requirements.txt
attrs==19.3.0
Django==2.2.15
django-rest-framework==0.1.0
djangorestframework==3.11.1
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
